# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Network Clone",
        "description": "Social Network Clone API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "feed/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get user videos.",
                "operationId": "ce39be3c0c8a0076c6a892fcba352490",
                "responses": {
                    "200": {
                        "description": "User videos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "feed/posts/{post}/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get videos for a post.",
                "operationId": "458d5f7440575bfadb1c7877afab5623",
                "responses": {
                    "200": {
                        "description": "Post videos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "Add a video to a post.",
                "operationId": "53bcc01ad906cb3a83049ea481128763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "video_url"
                                ],
                                "properties": {
                                    "video_url": {
                                        "type": "string",
                                        "example": "https://www.youtube.com/watch?v=12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Video added",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "feed/posts/{post}/videos/{video}": {
            "put": {
                "tags": [
                    "Videos"
                ],
                "summary": "Update a video for a post.",
                "operationId": "513a72e2722ac3e34a409f5192bb557f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "video_url"
                                ],
                                "properties": {
                                    "video_url": {
                                        "type": "string",
                                        "example": "https://www.youtube.com/watch?v=12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Video updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Videos"
                ],
                "summary": "Delete a video from a post.",
                "operationId": "cfd975ee39b1260fd68e81f7a333eed7",
                "responses": {
                    "200": {
                        "description": "Video deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
...
}